# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "RealmForge Control API",
    "description": "Governance kernel REST API — session, command, audit, snapshot, rollback, boards, and more.",
    "license": {
      "name": "UNLICENSED",
      "identifier": "UNLICENSED"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/actors/{id}/scope": {
      "get": {
        "tags": [
          "actors"
        ],
        "summary": "Get the full ActorScope for an actor and session.",
        "operationId": "get_actor_scope",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Actor ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "session_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "ses_abc123"
          }
        ],
        "responses": {
          "200": {
            "description": "Actor scope returned",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActorScopeResponse"
                }
              }
            }
          },
          "404": {
            "description": "Actor or session not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/v1/audit/chain/verify": {
      "get": {
        "tags": [
          "audit"
        ],
        "summary": "Verify the integrity of the full audit hash chain for a project.",
        "operationId": "verify_audit_chain",
        "parameters": [
          {
            "name": "project_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Chain verification result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChainVerifyResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid project ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/v1/audit/events": {
      "get": {
        "tags": [
          "audit"
        ],
        "summary": "Query audit events for a project with optional filters and pagination.",
        "operationId": "query_audit_events",
        "parameters": [
          {
            "name": "project_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "event_type",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "actor_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "entity_type",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "from_time",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ],
              "format": "date-time"
            }
          },
          {
            "name": "to_time",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ],
              "format": "date-time"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int64",
              "minimum": 0
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int64",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Events returned",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventsListResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid query parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/v1/audit/stream": {
      "get": {
        "tags": [
          "audit"
        ],
        "summary": "Stream audit events for a project as Server-Sent Events (SSE).",
        "description": "SSE endpoint — streams audit events for a project as they arrive.\nReconnect-safe: pass the `Last-Event-ID` header with the last received\nevent's offset to resume without gaps.",
        "operationId": "stream_audit_events",
        "parameters": [
          {
            "name": "project_id",
            "in": "query",
            "description": "Project ID to stream events for",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Last-Event-ID",
            "in": "header",
            "description": "Resume offset from reconnect",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "SSE stream of audit events (text/event-stream)"
          },
          "400": {
            "description": "Invalid project ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ActorScopeResponse": {
        "type": "object",
        "required": [
          "scope"
        ],
        "properties": {
          "scope": {
            "type": "object"
          }
        }
      },
...
      "ChainVerifyResponse": {
        "type": "object",
        "required": [
          "chain_integrity",
          "event_count"
        ],
        "properties": {
          "chain_integrity": {
            "type": "boolean"
          },
          "event_count": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          }
        }
      },
...
      "EventsListResponse": {
        "type": "object",
        "required": [
          "events",
          "total"
        ],
        "properties": {
          "events": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "total": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          }
        }
      },
...
      "ProblemDetails": {
        "type": "object",
        "description": "RFC 7807 Problem Details response, extended with a machine-readable error code\nand trace ID per the RealmForge API error standard.",
        "required": [
          "code",
          "title",
          "status",
          "detail"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "Machine-readable SCREAMING_SNAKE_CASE error code. Stable across versions."
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "title": {
            "type": "string"
          },
          "trace_id": {
            "type": [
              "string",
              "null"
            ],
            "description": "Correlation ID matching the X-Request-Id response header."
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "session_token"
      }
    }
  },
...
}